requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=8.3.0
tqdm>=4.62.0
tensorboard>=2.8.0
pyyaml>=6.0
scikit-learn>=1.0.0
jupyter>=1.0.0
notebook>=6.0.0